move to template_func
